gzilla/requirements.txt
pyyaml
scapy
